Refactor: Replace Utils references with Constants and remove unused Utils class
