+++ FurLab.CLI/Commands/Database/PgPass/Add/PgPassAddCommand.cs
@@ -1,6 +1,6 @@
 using FurLab.Core.Interfaces;
 using FurLab.Core.Models;
-using FurLab.Core.Utils;
+using FurLab.Core.Constants;
 using Spectre.Console;
 using Spectre.Console.Cli;
